Ref: Move common widgets -> lwtui.widget
